Latihan 6 Feb 2023

+++ application/controllers/Farmasi_po.php
@@ -86,10 +86,11 @@ class Farmasi_po extends CI_Controller
 
 	public function ajax_list($param)
 	{
-		$lvx = $this->session->userdata('user_level');
-		$level = $this->session->userdata('level');
-		$akses = $this->M_global->cek_menu_akses($level, 3101);
-		$dat   = explode("~", $param);
+		$lvx    = $this->session->userdata('user_level');
+		$level  = $this->session->userdata('level');
+		$akses  = $this->M_global->cek_menu_akses($level, 3101);
+		$lock   = $this->M_global->close_app();
+		$dat    = explode("~", $param);
 		if ($dat[0] == 1) {
 			$bulan = date('m');
 			$tahun = date('Y');
@@ -125,7 +126,7 @@ class Farmasi_po extends CI_Controller
 			$row[] = '<span "font-weight:bold;"><b>' . $rd->nama . '</b></span>';
 			$row[] = '<span "font-weight:bold;"><b>' . $status . '</b></span>';
 
-			if ($akses->uedit == 1 && $akses->udel == 1) {
+			if ($akses->uedit == 1 && $akses->udel == 1 && $lock == 0) {
 				$email = $this->db->get_where('tbl_vendor', ['vendor_id' => $rd->vendor_id])->row_array();
 				if ($rd->setuju == 0) {
 					if ($lvx >= 2) {
@@ -158,16 +159,12 @@ class Farmasi_po extends CI_Controller
 							<a class="btn btn-sm btn-primary" href="' . base_url("farmasi_po/show/" . $rd->id . "") . '" title="Show" >
 								<i class="glyphicon glyphicon-eye-open"></i> 
 							</a>
-							
-							<a class="btn btn-sm btn-warning" onclick="javascript:cekctk(' . "'" . $rd->id . "'" . ",'" . $rd->po_no . "'" . ');" title="Cetak1" >
-								<i class="glyphicon glyphicon-print"></i>
-							</a>
 						';
 					}
 				}
-			} else if ($akses->uedit == 1 && $akses->udel == 0) {
+			} else if ($akses->uedit == 1 && $akses->udel == 0 && $lock == 0) {
 				$row[] = '<a class="btn btn-sm btn-primary" href="' . base_url("farmasi_po/edit/" . $rd->id . "") . '" title="Edit" ><i class="glyphicon glyphicon-edit"></i> </a> ';
-			} else if ($akses->uedit == 0 && $akses->udel == 1) {
+			} else if ($akses->uedit == 0 && $akses->udel == 1 && $lock == 0) {
 				$row[] = '<a class="btn btn-sm btn-danger" href="javascript:void(0)" title="Hapus" onclick="delete_data(' . "'" . $rd->id . "'" . ')"><i class="glyphicon glyphicon-trash"></i> </a>';
 			} else {
 				$row[] = '';
@@ -400,6 +397,7 @@ class Farmasi_po extends CI_Controller
 			'tglkirim'      => $this->input->post('tanggalkirim'),
 			'vendor_id'     => $this->input->post('supp'),
 			'username'      => $userid,
+			'disetujuioleh'  => $userid,
 			'kurs'          => $this->input->post('kurs'),
 			'kursrate'      => $this->input->post('rate'),
 			'dikirimvia'    => $this->input->post('dikirimvia'),
@@ -420,21 +418,21 @@ class Farmasi_po extends CI_Controller
 		$qty    = $this->input->post('qty');
 		$sat    = $this->input->post('sat');
 		$harga  = $this->input->post('harga');
-		$disc   = $this->input->post('disc');
-		$tax    = $this->input->post('tax');
+		// $disc   = $this->input->post('disc');
+		// $tax    = $this->input->post('tax');
 		$jumlah = $this->input->post('jumlah');
 
 		$jumdata    = count($kode);
 		for ($i = 0; $i <= $jumdata - 1; $i++) {
 			$_harga    = str_replace(',', '', $harga[$i]);
-			$_disc     = str_replace(',', '', $disc[$i]);
+			// $_disc     = str_replace(',', '', $disc[$i]);
 			$_jumlah   = str_replace(',', '', $jumlah[$i]);
 
-			if ($tax[$i]) {
-				$_tax = 1;
-			} else {
-				$_tax = 0;
-			}
+			// if ($tax[$i]) {
+			// 	$_tax = 1;
+			// } else {
+			// 	$_tax = 0;
+			// }
 
 			$data_rinci = array(
 				'koders'       => $cabang,
@@ -442,9 +440,9 @@ class Farmasi_po extends CI_Controller
 				'kodebarang'   => $kode[$i],
 				'qty_po'       => $qty[$i],
 				'price_po'     => $_harga,
-				'discount'     => $_disc,
+				'discount'     => 0,
 				'satuan'       => $sat[$i],
-				'vat'          => $_tax,
+				'vat'          => 0,
 				'total'        => $_jumlah,
 
 			);
